[Mobile] Convert QMobileUI to swift 4 #90739
Imported from Git
 Author: Eric Marchand <eric.marchand@4d.com> 1512138272 +0100
 Committer: Eric Marchand <eric.marchand@4d.com> 1512138272 +0100
 sha1: cf22b2aef1f228a40ae52dfb4051f548e8cf02a8
 push-state: complete
 parent-changes: 730fca5d8316e4789977332a5ca2d5df7b752709=[217279]

[git-p4: depot-paths = "//depot/4eDimension/main/QMobile/QMobileUI/": change = 217976]

+++ Sources/Extensions/DateFormatter+QMobile.swift
@@ -116,7 +116,7 @@ open class TimeFormatter {
     }
 
     open func string(from number: NSNumber) -> String {
-        return string(from: TimeInterval(number))
+        return string(from: TimeInterval(number.doubleValue))
     }
 
     open func time(from string: String) -> TimeInterval? {
